refactor: move local-assets into root of config-loader and publicly expose it
